requirements.txt
django
python-social-auth
gspread
reportlab
django-choices
beautifulsoup4
lxml
django-sortedm2m